Upg OrientDB #3: Upgrade OrientDB deps to version 2.2.22
This commit upgrades OrientDB dependencies to the latest version.
Several alterations to the repo service has to be done, namely:
* OrientDB does not accept map keys with "." (dot) and "[" characters as
  a ODocument field name. These So escaping of these characters must be
  used while performing conversion to ODocument
* "factory__pid" map key has value of type TargetedID which is not
  serializable and cannot be stored as a field value in OrientDB.
* Fix unbalanced parentheses in SQL syntax that was silently ignored in
  1.7.x but treated as an error in 2.x

This commit also removes dependency on snappy-java since it does not
work on non-glibc based Linux systems. This will may cause problems in
OrienDB based deployments that have explicitly enabled compression
support in OrientDB.

Please note that proper function of OrientDB Studio has not been tested
so it is possible it does not work for this version of OrientDB.